Check and Install R Package

This function checks if an R package is installed, and if not, attempts to install it using either the standard CRAN repository or the Bioconductor repository.

Usage

check_install(package_name)

Arguments

package_name

A character string specifying the name of the package to be checked and installed.

Value

TRUE if the package is already installed or can be installed. FALSE otherwise

Examples

# \donttest{
# Check and install "dplyr" package
check_install("dplyr")
#> Loading required package: dplyr
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
# }