extract package names and install them

This function extracts all package names that are needed to run the code returned by the agent and installs them as needed.

Usage

extractInstallPkg(code)

Arguments

code

code block returned by the agent.

Value

final status of the packages as a logical vector, TRUE if packages is already installed or could be installed. FALSE if the package can't be install.

Examples

# \donttest{
# Check code for packages that need installing
code <- "library(devtools)\n x<-5"
extractInstallPkg(code)
#> Loading required package: devtools
#> Warning: there is no package called ‘devtools’
#> 
#> trying to install: devtools
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> 'help("repositories", package = "BiocManager")' for details.
#> Replacement repositories:
#>     CRAN: https://cran.rstudio.com
#> Bioconductor version 3.19 (BiocManager 1.30.25), R 4.4.1 (2024-06-14)
#> Installing package(s) 'devtools'
#> also installing the dependencies ‘credentials’, ‘zip’, ‘xtable’, ‘sourcetools’, ‘later’, ‘Rcpp’, ‘selectr’, ‘clipr’, ‘gert’, ‘shiny’, ‘htmlwidgets’, ‘xopen’, ‘brew’, ‘commonmark’, ‘httpuv’, ‘crosstalk’, ‘promises’, ‘rvest’, ‘gitcreds’, ‘ini’, ‘gargle’, ‘cyclocomp’, ‘xmlparsedata’, ‘whoami’, ‘hunspell’, ‘usethis’, ‘ellipsis’, ‘miniUI’, ‘profvis’, ‘rcmdcheck’, ‘remotes’, ‘roxygen2’, ‘rversions’, ‘urlchecker’, ‘DT’, ‘foghorn’, ‘gh’, ‘gmailr’, ‘lintr’, ‘mockery’, ‘pingr’, ‘rhub’, ‘rstudioapi’, ‘spelling’
#> Loading required package: devtools
#> Loading required package: usethis
#> devtools 
#>     TRUE 
# }