Send a prompt to a specified language model agent and return the response.

Send a prompt to a specified language model agent and return the response.

Usage

sendPrompt(
  agent,
  prompt,
  context = rbionfoExp,
  return.type = c("text", "object"),
  ...
)

Arguments

agent

An object containing the agent's information (e.g., type and model etc.).

prompt

The prompt text to send to the language model.

context

Optional context to provide alongside the prompt (default is rbionfoExp).

return.type

The type of output to return, either the text response ("text") or the entire response object ("object").

...

Additional arguments to be passed to the prompt function.

Value

The text response or the entire response object, based on the specified return type.

See also

Examples

if (FALSE) {
agent <- setupAgent(name="openai",type="chat",model="gpt-4",
                    ai_api_key=Sys.getenv("OPENAI_API_KEY"))
prompt <- "tell me a joke"
response <- sendPrompt(agent, prompt)

response <- sendPrompt(agent,prompt,context=rbionfoExp,return.type="text", max_tokens = 500)
}